Add support for "proper" tail recursion
Both branches of the `if` function are now returned to the caller
unevaluated and the caller of the function then evaluates them
instead. This means that a tail call to another function replaces the
frame of the caller with the to-be-called function, resulting in proper
tail recursion. The eval_cell() function also does the same and
evaluates the function body in the same frame without going through
eval() first.