Support tailcalls in progn
The last element of progn is now evaluated in a "tail context". This
makes it possible to implement some sorts of loops with it.

Also modified the tests to run with the default 1MiB heap memory
limit. This should test the GC more and with tail recursion in place,
the tests no longer run out of memory.