Use unique bits in ObjectType
There's exactly 8 types that need to be handled in a separate
manner. The Nil, True and TailCall constants can share a single type as
the type itself is never compared. These wouldn't technically need a
type but it's nice to still have one, even if it's shared between the
two constants.