Clean up Game of Life example
Reorganized the code a bit, added a hard stop to the simulation and
created a function that generates the initial state that is then called
before the main loop.

The standard functions are also now compiled. This both improves the
performance as well as avoids accidental overwriting of the _impl
functions by users.

Also added a make target for the demo that can be run with `make demo`.