Lower default memory high water to 70%
The old default of 90% caused garbage collection to happen very
frequently when the memory usage went above a certain value. The
temporary values that were created during the various computations
triggered garbage collection and most of the time went into copying the
existing objects.

The new default of 70% should leave more memory in place for temporary
objects so that garbage collection happens less often. Even this is not
ideal for the Game of Life demo as it seems to perform best when there's
plenty of free memory around, e.g. with `-m 30`.