Use tagged pointers for objects
This removes the need to allocate memory for numbers which cuts down on
the memory usage. It also simplifies equality comparisons to a simple
pointer comparison and the less than comparison can ignore the type
completely as it only makes sense for numbers.

The implementation relies on two's complement representation for
integers which will be true for x86_64 that the compiler will target.

Also split the code into a source file and a header. This'll make it
easier to split out the compilation part into its own file.