Add compiler skeleton
The old compile function only resolved the symbols that weren't formal
parameters to their values. The end result is that the function will
effectively be compiled into a pure function that can refer to variables
outside of the scope but the values for them will be frozen to their
values at "compile time". This behavior is still accessible under the
"freeze" function. The compilation of it will in the future actually
turn the functions into bytecode, if possible.