Add liveness analysis and register allocation
The liveness analysis is based on the assumption that when the variable
is first assigned into, it will not pass itself into the set of outgoing
live variables. This is simple as each variable is only assigned to once
and the variables aren't used across branches.

The register allocation is not very smart as values get read more than
once if they're referred to by the conditional or the branches of an if
statement.