Add initial Bite compilation
The bite-code for constants, variables and addition is now compiled into
machine code. The register allocation scheme uses the Sethi-Ullman
algorithm with a simplification that left leaf nodes are always stored
in a register, even if they are constants and could otherwise be
operated on without a register. This also helps with the case where
64-bit immediate values are used as they can't be added directly.