Compile if bites
The CMP instruction used the wrong opcode and it compared the values in
the inverse order. The old direct compilation mode had the arguments
evaluated in the "correct" order which is why it seemed to work even if
the documented behavior was different. Swapping the order of the
arguments fixes this problem.

The if bite compilation has a small optimization where it uses the
result of a OP_EQ or OP_LESS bite directly without storing the nil or
true values in a register.