Add separate UserFunction type
This makes the size calculations easier for the memory allocation.

Moved the JIT compiled memory into its own variable instead of reusing
the function body. This way the original body is not lost after
compilation and could potentially be printed. However, printing of
self-recursive functions does not yet work.

Made the debug option work the way that it usually does. Using the -d
flag multiple times increases the debug level. Removed the separate
stacktrace option and made the default debug level less verbose.

Improved the debug output to include the name of the symbol that points
to a builtin or user-defined function.