Add support for the cons builtin
The cons builtin is now supported but the garbage collection doesn't
support it yet. In order for this to be supported, a separate stack is
needed for the JIT objects as the current approach stores both the
object pointers as well as actual execution related data like pushed
registers. To be able to store the data separately so that the pointers
can be rewritten by the garbage collection, they can be stored in a
separate stack.