Use a custom stack for JIT code
Using the actual stack of the program is error prone as controlling what
gets placed on it and in what alignment is too hard for it to be a
reliable way to deal with the problem of tracking the JIT stack. Thus,
the RSI register has to be sacrified to store a pointer to a
stack-allocated array of Object pointers.