Immediately initialize reserved stack
Since the compilation of a bite can now always trigger a garbage
collection to take place, any stack space that is used must be
immediately initialized. Otherwise the garbage collection will scan the
stack and find only garbage in the memory.

This also solves the limitation of only 8 function arguments as they are
now pushed one at a time to the stack.